move SubtitleBox to components/
