+++ src/components/subtitle/SubtitleBox.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
 import { useTranslation } from 'react-i18next';
-import Subtitle from '../components/subtitle/Subtitle';
-import subtitlesData from '../config/i18n/ko/Subtitles.json';
+import Subtitle from './Subtitle';
+import subtitlesData from '../../config/i18n/ko/Subtitles.json';
 
 const getSubtitleKey = () => {
   const key = Object.keys(subtitlesData)
